requirements.txt
flask==3.0.0
opencv-python==4.8.0
pillow==10.0.0
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0 